pyproject.toml
[project]
name = "audithound"
version = "0.1.0"
description = "A robust TUI application for security audit compliance scanning"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Team", email = "team@example.com"}
]
license = {file = "LICENSE"}
keywords = ["security", "audit", "compliance", "sast", "secrets"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "docker>=7.1.0",
    "jsonschema>=4.25.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "subprocess-tee>=0.4.2",
    "textual>=6.1.0",
    "toml>=0.10.2",
    "typer>=0.17.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
scanners = [
    "bandit>=1.8.6",
    "trufflehog",
    "semgrep", 
    "safety",
    "checkov",
]

[project.scripts]
audithound = "audithound.main:app"

[project.urls]
Homepage = "https://github.com/pesnik/audithound"
Repository = "https://github.com/pesnik/audithound"
Issues = "https://github.com/pesnik/audithound/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "B", "C", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
